Folder refactor done. App is working. Heading towards ViewModel classes refactor
