--- app/src/main/java/com/practicum/playlistmaker/data/dto/NetworkResponse.kt
@@ -1,8 +1,0 @@
-package com.practicum.playlistmaker.data.dto
-
-import okhttp3.Request
-
-open class NetworkResponse {
-    var resultCode = 0
-    var requestInfo: Request? = null
-}
