--- app/src/main/java/com/practicum/playlistmaker/data/dto/TrackDto.kt
@@ -1,15 +1,0 @@
-package com.practicum.playlistmaker.data.dto
-
-class TrackDto(
-    val trackId: String,
-    val artistName: String,
-    val trackName: String,
-    val trackTimeMillis: Long,
-    val previewUrl: String,
-    val artworkUrl60: String,
-    val artworkUrl100: String,
-    val collectionName: String,
-    val releaseDate: String,
-    val primaryGenreName: String,
-    val country: String,
-)
