--- app/src/main/java/com/practicum/playlistmaker/data/dto/TrackListRequest.kt
@@ -1,3 +1,0 @@
-package com.practicum.playlistmaker.data.dto
-
-data class TrackListRequest(val expression: String)
