--- app/src/main/java/com/practicum/playlistmaker/data/dto/TrackListResponse.kt
@@ -1,6 +1,0 @@
-package com.practicum.playlistmaker.data.dto
-
-class TrackListResponse(
-    val resultCount: Int,
-    val results: ArrayList<TrackDto>
-): NetworkResponse()
