--- app/src/main/java/com/practicum/playlistmaker/data/network/ItunesAPI.kt
@@ -1,11 +1,0 @@
-package com.practicum.playlistmaker.data.network
-
-import com.practicum.playlistmaker.data.dto.TrackListResponse
-import retrofit2.Call
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface ItunesAPI {
-    @GET("/search?entity=song")
-    fun getSongsList(@Query("term") artistName : String) : Call<TrackListResponse>
-}
