--- app/src/main/java/com/practicum/playlistmaker/data/repository/TracklistNetworkClient.kt
@@ -1,9 +1,0 @@
-package com.practicum.playlistmaker.data.repository
-
-import com.practicum.playlistmaker.data.dto.NetworkResponse
-import com.practicum.playlistmaker.data.dto.TrackListResponse
-import retrofit2.Callback
-
-interface TracklistNetworkClient {
-    fun doTrackRequest(dto: Any?) : NetworkResponse
-}
