--- app/src/main/java/com/practicum/playlistmaker/data/storage/SharedPrefsClient.kt
@@ -1,13 +1,0 @@
-package com.practicum.playlistmaker.data.storage
-
-import android.content.Context
-import android.content.SharedPreferences
-
-
-class SharedPrefsClient(private val context: Context, private val key: String) {
-
-    fun getSharedPrefs(): SharedPreferences {
-        val sharedPrefs = context.getSharedPreferences(key, Context.MODE_PRIVATE)
-        return sharedPrefs
-    }
-}
