--- app/src/main/java/com/practicum/playlistmaker/domain/consumer/Consumer.kt
@@ -1,5 +1,0 @@
-package com.practicum.playlistmaker.domain.consumer
-
-interface Consumer<T> {
-    fun consume(data: ConsumerData<T>)
-}
