--- app/src/main/java/com/practicum/playlistmaker/domain/consumer/ConsumerData.kt
@@ -1,7 +1,0 @@
-package com.practicum.playlistmaker.domain.consumer
-
-sealed interface ConsumerData<T> {
-    data class Data<T>(val value: T?) :ConsumerData<T>
-    data class Error<T>(val message: String) :ConsumerData<T>
-    data class NoConnection<T>(val message: String) :ConsumerData<T>
-}
