--- app/src/main/java/com/practicum/playlistmaker/domain/repository/AppThemeRepository.kt
@@ -1,6 +1,0 @@
-package com.practicum.playlistmaker.domain.repository
-
-interface AppThemeRepository {
-    fun setDarkTheme(isDarkTheme: Boolean)
-    fun checkIsDarkThemeOn(): Boolean
-}
