--- app/src/main/java/com/practicum/playlistmaker/player/domain/mapper/TrackInfoMapper.kt
@@ -1,23 +1,0 @@
-package com.practicum.playlistmaker.player.domain.mapper
-
-import com.practicum.playlistmaker.search.domain.track_model.Track
-import com.practicum.playlistmaker.player.domain.model.TrackInfoModel
-
-// Аналогично, Маппер для непригодившихся сейчас моделей. Оставил на будущее.
-
-object TrackInfoMapper {
-    fun map (track: Track): TrackInfoModel {
-        return TrackInfoModel(
-            trackId = track.trackId,
-            trackName = track.trackName,
-            artistName = track.artistName,
-            trackTimeMillis = track.trackTimeMillis,
-            previewUrl = track.previewUrl,
-            artworkUrl100 = track.artworkUrl100,
-            collectionName = track.collectionName,
-            releaseDate = track.releaseDate,
-            primaryGenreName = track.primaryGenreName,
-            country = track.country
-        )
-    }
-}
