+++ app/src/main/java/com/practicum/playlistmaker/player/presentation/mapper/DateFormatter.kt
@@ -1,0 +1,9 @@
+package com.practicum.playlistmaker.player.presentation.mapper
+
+import java.text.SimpleDateFormat
+import java.util.Locale
+
+object DateFormatter {
+    val timeFormatter = SimpleDateFormat("mm:ss", Locale.getDefault())
+    val yearFormatter = SimpleDateFormat("yyyy", Locale.getDefault())
+}
