+++ app/src/main/java/com/practicum/playlistmaker/player/presentation/model/TrackSearchModel.kt
@@ -1,0 +1,12 @@
+package com.practicum.playlistmaker.player.presentation.model
+
+
+// Этой модели не нашел применение( Буду очень рад любым комментариям!
+data class TrackSearchModel(
+    val trackId: String,
+    val artistName: String,
+    val trackName: String,
+    val trackTimeMillis: Long,
+    val artworkUrl60: String,
+) {
+}
