+++ app/src/main/java/com/practicum/playlistmaker/search/data/dto/TrackListRequest.kt
@@ -1,0 +1,3 @@
+package com.practicum.playlistmaker.search.data.dto
+
+data class TrackListRequest(val expression: String)
