+++ app/src/main/java/com/practicum/playlistmaker/search/data/dto/TrackListResponse.kt
@@ -1,0 +1,6 @@
+package com.practicum.playlistmaker.search.data.dto
+
+class TrackListResponse(
+    val resultCount: Int,
+    val results: ArrayList<TrackDto>
+): NetworkResponse()
