+++ app/src/main/java/com/practicum/playlistmaker/search/data/mapper/TrackMapper.kt
@@ -1,0 +1,22 @@
+package com.practicum.playlistmaker.search.data.mapper
+
+import com.practicum.playlistmaker.search.data.dto.TrackDto
+import com.practicum.playlistmaker.search.domain.track_model.Track
+
+object TrackMapper {
+    fun map (trackDto: TrackDto): Track {
+        return Track(
+            trackId = trackDto.trackId,
+            trackName = trackDto.trackName,
+            artistName = trackDto.artistName,
+            trackTimeMillis = trackDto.trackTimeMillis,
+            previewUrl = trackDto.previewUrl,
+            artworkUrl60 = trackDto.artworkUrl60,
+            artworkUrl100 = trackDto.artworkUrl100,
+            collectionName = trackDto.collectionName,
+            releaseDate = trackDto.releaseDate,
+            primaryGenreName = trackDto.primaryGenreName,
+            country = trackDto.country
+        )
+    }
+}
