+++ app/src/main/java/com/practicum/playlistmaker/search/data/network/ItunesAPI.kt
@@ -1,0 +1,11 @@
+package com.practicum.playlistmaker.search.data.network
+
+import com.practicum.playlistmaker.search.data.dto.TrackListResponse
+import retrofit2.Call
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface ItunesAPI {
+    @GET("/search?entity=song")
+    fun getSongsList(@Query("term") artistName : String) : Call<TrackListResponse>
+}
