+++ app/src/main/java/com/practicum/playlistmaker/search/data/repository/TracklistNetworkClient.kt
@@ -1,0 +1,7 @@
+package com.practicum.playlistmaker.search.data.repository
+
+import com.practicum.playlistmaker.search.data.dto.NetworkResponse
+
+interface TracklistNetworkClient {
+    fun doTrackRequest(dto: Any?) : NetworkResponse
+}
