+++ app/src/main/java/com/practicum/playlistmaker/search/domain/tracks_intr/GetTrackListFromServerUseCase.kt
@@ -1,7 +1,7 @@
 package com.practicum.playlistmaker.search.domain.tracks_intr
 
-import com.practicum.playlistmaker.domain.consumer.Consumer
-import com.practicum.playlistmaker.domain.consumer.ConsumerData
+import com.practicum.playlistmaker.search.domain.consumer.Consumer
+import com.practicum.playlistmaker.search.domain.consumer.ConsumerData
 import com.practicum.playlistmaker.search.domain.track_model.Resourse
 import com.practicum.playlistmaker.search.domain.track_model.Track
 import com.practicum.playlistmaker.search.domain.repository.TrackListRepository
