+++ app/src/main/java/com/practicum/playlistmaker/settings/domain/repository/AppThemeRepository.kt
@@ -1,0 +1,6 @@
+package com.practicum.playlistmaker.settings.domain.repository
+
+interface AppThemeRepository {
+    fun setDarkTheme(isDarkTheme: Boolean)
+    fun checkIsDarkThemeOn(): Boolean
+}
